Исправила замечания ревьюера
