+++ api_yamdb/users/models.py
@@ -1,6 +1,5 @@
 from django.contrib.auth.models import AbstractUser
 from django.db import models
-from django.utils.translation import gettext_lazy as _
 
 from reviews.constants import MAX_LENGTH_CODE, MAX_LENGTH_PSW, MAX_LENGTH_ROLE
 
@@ -27,7 +26,7 @@ class User(AbstractUser):
         verbose_name='Код подтверждения'
     )
     password = models.CharField(
-        _('password'), max_length=MAX_LENGTH_PSW, blank=True, null=True
+        'password', max_length=MAX_LENGTH_PSW, blank=True, null=True
     )
 
     class Meta:
